+++ backend/src/main/java/coursepick/coursepick/application/ReportAlerter.java
@@ -1,9 +1,11 @@
 package coursepick.coursepick.application;
 
 import coursepick.coursepick.domain.course.Course;
+import coursepick.coursepick.domain.course.Review;
 
 public interface CourseReportAlerter {
 
     void alert(Course course);
 
+    void alert(Course course, Review review);
 }
